Restructure project towards a traditional python package
Eventually make all executables subcommands of `classify`. Disperse functionality from top-level
executable scripts like metagenomics.py and taxon_filter.py into the `classify` library.

Restructure tests to remove `unit/` folder and place unit integration tests together for each file,
with pytest marks to distinguish between unit and integration tests.
