+++ .travis.yml
@@ -63,7 +63,7 @@ jobs:
         - docker pull $DOCKER_TAG
         - mkdir coverage
       script:
-        - docker run -e _JAVA_OPTIONS -e PYTEST_ADDOPTS -v `pwd`/test:/opt/viral-ngs/source/test -v `pwd`/coverage:/coverage --entrypoint /bin/bash $DOCKER_TAG -c 'set -e; cd /opt/viral-ngs/source; pytest test/unit; cp .coverage /coverage'
+        - docker run -e _JAVA_OPTIONS -e PYTEST_ADDOPTS -v `pwd`/test:/opt/viral-ngs/source/test -v `pwd`/coverage:/coverage --entrypoint /bin/bash $DOCKER_TAG -c 'set -e; cd /opt/viral-ngs/source; pytest -m "not integration" test; cp .coverage /coverage'
       after_success:
         - mv coverage/.coverage .
         - pip install coveralls==1.1
