+++ Dockerfile
@@ -13,11 +13,8 @@ RUN $VIRAL_NGS_PATH/docker/install-conda-dependencies.sh $VIRAL_CLASSIFY_PATH/re
 COPY . $VIRAL_CLASSIFY_PATH
 
 # Link key bits of python code into the path
-RUN ln -s $VIRAL_CLASSIFY_PATH/metagenomics.py $VIRAL_CLASSIFY_PATH/taxon_filter.py $VIRAL_CLASSIFY_PATH/kmer_utils.py $VIRAL_CLASSIFY_PATH/classify $VIRAL_NGS_PATH
+RUN ln -s $VIRAL_CLASSIFY_PATH/classify/metagenomics.py $VIRAL_CLASSIFY_PATH/classify/taxon_filter.py $VIRAL_CLASSIFY_PATH/classify/kmer_utils.py $VIRAL_CLASSIFY_PATH/classify $VIRAL_NGS_PATH
 
-# This not only prints the current version string, but it
-# also saves it to the VERSION file for later use and also
-# verifies that conda-installed python libraries are working.
-RUN /bin/bash -c "set -e; echo -n 'viral-ngs version: '; metagenomics.py --version"
+RUN /bin/bash -c "set -e; echo -n 'viral-ngs version: ' "
 
 CMD ["/bin/bash"]
