+++ classify/blast.py
@@ -1,17 +1,76 @@
 "Tools in the blast+ suite."
-
+import collections
+import itertools
 import logging
+import operator
 import os
+from pathlib import Path
 import subprocess
+
+import ncbitax
 
 import tools
 import tools.samtools
 import util.misc
+import classify.metagenomics
 
 TOOL_NAME = "blast"
 TOOL_VERSION = "2.7.1"
 
 _log = logging.getLogger(__name__)
+
+
+def process_blast_hits(db, hits, top_percent):
+    '''Filter groups of blast hits and perform lca.
+
+    Args:
+      db: (TaxonomyDb) Taxonomy db.
+      hits: []BlastRecord groups of hits.
+      top_percent: (float) Only consider hits within this percent of top bit score.
+
+    Return:
+      (int) Tax id of LCA.
+    '''
+    hits = (translate_gi_to_tax_id(db, hit) for hit in hits)
+
+    hits = [hit for hit in hits if hit.subject_id != 0]
+    if len(hits) == 0:
+        return
+
+    best_score = max(hit.bit_score for hit in hits)
+    cutoff_bit_score = (100 - top_percent) / 100 * best_score
+    valid_hits = (hit for hit in hits if hit.bit_score >= cutoff_bit_score)
+    valid_hits = list(valid_hits)
+    # Sort requires realized list
+    valid_hits.sort(key=operator.attrgetter('bit_score'), reverse=True)
+    if valid_hits:
+        tax_ids = tuple(itertools.chain(*(blast_m8_taxids(hit) for hit in valid_hits)))
+        return ncbitax.coverage_lca(tax_ids, db.parents)
+
+
+def translate_gi_to_tax_id(db, record):
+    '''Replace gi headers in subject ids to int taxonomy ids.'''
+    gi = int(record.subject_id.split('|')[1])
+    tax_id = db.gis[gi]
+    rec_list = list(record)
+    rec_list[1] = tax_id
+    return BlastRecord(*rec_list)
+
+
+def paired_query_id(record):
+    '''Replace paired suffixes in query ids.'''
+    suffixes = ('/1', '/2')
+    for suffix in suffixes:
+        if record.query_id.endswith(suffix):
+            rec_list = list(record)
+            rec_list[0] = record.query_id[:-len(suffix)]
+            return BlastRecord(*rec_list)
+    return record
+
+
+def blast_m8_taxids(record):
+    return [int(record.subject_id)]
+
 
 class BlastTools(tools.Tool):
     """'Abstract' base class for tools in the blast+ suite.
@@ -93,7 +152,7 @@ class MakeblastdbTool(BlastTools):
         # or a list of strings
         if 'basestring' not in globals():
            basestring = str
-        if isinstance(fasta_files, basestring):
+        if isinstance(fasta_files, basestring) or isinstance(fasta_files, Path):
             fasta_files = [fasta_files]
         elif isinstance(fasta_files, list):
             pass
@@ -114,3 +173,64 @@ class MakeblastdbTool(BlastTools):
         self.execute(*args)
 
         return database_prefix_path
+
+
+BlastRecord = collections.namedtuple(
+    'BlastRecord', [
+        'query_id', 'subject_id', 'percent_identity', 'aln_length', 'mismatch_count', 'gap_open_count', 'query_start',
+        'query_end', 'subject_start', 'subject_end', 'e_val', 'bit_score', 'extra'
+    ]
+)
+
+
+def blast_records(f):
+    '''Yield blast m8 records line by line'''
+    for line in f:
+        if line.startswith('#'):
+            continue
+        parts = line.strip().split()
+        for field in range(3, 10):
+            parts[field] = int(parts[field])
+        for field in (2, 10, 11):
+            parts[field] = float(parts[field])
+        args = parts[:12]
+        extra = parts[12:]
+        args.append(extra)
+
+        yield BlastRecord(*args)
+
+
+def blast_lca(db,
+              m8_file,
+              output,
+              paired=False,
+              min_bit_score=50,
+              max_expected_value=0.01,
+              top_percent=10,):
+    '''Calculate the LCA taxonomy id for groups of blast hits.
+
+    Writes tsv output: query_id \t tax_id
+
+    Args:
+      db: (TaxonomyDb) Taxonomy db.
+      m8_file: (io) Blast m8 file to read.
+      output: (io) Output file.
+      paired: (bool) Whether to count paired suffixes /1,/2 as one group.
+      min_bit_score: (float) Minimum bit score or discard.
+      max_expected_value: (float) Maximum e-val or discard.
+      top_percent: (float) Only this percent within top hit are used.
+    '''
+    records = blast_records(m8_file)
+    records = (r for r in records if r.e_val <= max_expected_value)
+    records = (r for r in records if r.bit_score >= min_bit_score)
+    if paired:
+        records = (paired_query_id(rec) for rec in records)
+    blast_groups = (v for k, v in itertools.groupby(records, operator.attrgetter('query_id')))
+    for blast_group in blast_groups:
+        blast_group = list(blast_group)
+        tax_id = process_blast_hits(db, blast_group, top_percent)
+        query_id = blast_group[0].query_id
+        if not tax_id:
+            log.debug('Query: {} has no valid taxonomy paths.'.format(query_id))
+        classified = 'C' if tax_id else 'U'
+        output.write('{}\t{}\t{}\n'.format(classified, query_id, tax_id))
