+++ classify/kraken.py
@@ -1,16 +1,18 @@
-'''
-KRAKEN metagenomics classifier
-'''
+import argparse
 import collections
 import itertools
 import logging
 import os
 import os.path
+import re
 import shlex
 import shutil
 import subprocess
 import sys
 import tempfile
+
+from Bio import SeqIO
+import ncbitax.subset
 
 import tools
 import tools.picard
@@ -24,6 +26,13 @@ KRAKENUNIQ_VERSION = '0.5.7_yesimon'
 
 
 log = logging.getLogger(__name__)
+
+
+class KrakenBuildError(Exception):
+    '''Error while building kraken database.'''
+
+class KrakenUniqBuildError(KrakenBuildError):
+    '''Error while building KrakenUniq database.'''
 
 
 class Kraken(tools.Tool):
@@ -273,12 +282,6 @@ class Kraken(tools.Tool):
                 subprocess.run(cmd, stdout=of, stderr=subprocess.PIPE, check=True)
 
 
-@tools.skip_install_test()
-class Jellyfish(Kraken):
-    """ Tool wrapper for Jellyfish (installed by kraken-all metapackage) """
-    subtool_name = 'jellyfish'
-
-
 class KrakenUniq(Kraken):
 
     BINS = {
@@ -376,3 +379,400 @@ class KrakenUniq(Kraken):
                 name = parts[8]
                 report[tax_id] = (tax_reads, tax_kmers)
         return report
+
+
+def parser_krakenuniq_build(parser=argparse.ArgumentParser()):
+    parser.add_argument('db', help='Krakenuniq database output directory.')
+    parser.add_argument('--library', help='Input library directory of fasta files. If not specified, it will be read from the "library" subdirectory of "db".')
+    parser.add_argument('--taxonomy', help='Taxonomy db directory. If not specified, it will be read from the "taxonomy" subdirectory of "db".')
+    parser.add_argument('--subsetTaxonomy', dest='subset_taxonomy', action='store_true', help='Subset taxonomy based on library fastas.')
+    parser.add_argument('--minimizerLen', dest='minimizer_len', type=int, help='Minimizer length (krakenuniq default: 15)')
+    parser.add_argument('--kmerLen', dest='kmer_len', type=int, help='k-mer length (krakenuniq default: 31)')
+    parser.add_argument('--maxDbSize', dest='max_db_size', type=int, help='Maximum db size in GB (will shrink if too big)')
+    parser.add_argument('--clean', action='store_true', help='Clean by deleting other database files after build')
+    parser.add_argument('--workOnDisk', dest='work_on_disk', action='store_true', help='Work on disk instead of RAM. This is generally much slower unless the "db" directory lives on a RAM disk.')
+    util.cmd.common_args(parser, (('threads', None), ('loglevel', None), ('version', None), ('tmp_dir', None)))
+    util.cmd.attach_main(parser, krakenuniq_build, split_args=True)
+    return parser
+def krakenuniq_build(db, library, taxonomy=None, subset_taxonomy=None,
+                     threads=None, work_on_disk=False,
+                     minimizer_len=None, kmer_len=None, max_db_size=None, clean=False):
+    '''
+    Builds a krakenuniq database from library directory of fastas and taxonomy
+    db directory. The --subsetTaxonomy option allows shrinking the taxonomy to
+    only include taxids associated with the library folders. For this to work,
+    the library fastas must have the standard accession id names such as
+    `>NC1234.1` or `>NC_01234.1`.
+
+    Setting the --minimizerLen (default: 16) small, such as 10, will drastically
+    shrink the db size for small inputs, which is useful for testing.
+
+    The built db may include symlinks to the original --library / --taxonomy
+    directories. If you want to build a static archiveable version of the
+    library, simply use the --clean option, which will also remove any
+    unnecessary files.
+    '''
+    util.file.mkdir_p(db)
+    library_dir = os.path.join(db, 'library')
+    library_exists = os.path.exists(library_dir)
+    if library:
+        try:
+            os.symlink(os.path.abspath(library), os.path.join(db, 'library'))
+        except FileExistsError:
+            pass
+    else:
+        if not library_exists:
+            raise FileNotFoundError('Library directory {} not found'.format(library_dir))
+
+    taxonomy_dir = os.path.join(db, 'taxonomy')
+    taxonomy_exists = os.path.exists(taxonomy_dir)
+    if taxonomy:
+        if taxonomy_exists:
+            raise KrakenUniqBuildError('Output db directory already contains taxonomy directory {}'.format(taxonomy_dir))
+        if subset_taxonomy:
+            accessions = fasta_library_accessions(library)
+
+            whitelist_accession_f = util.file.mkstempfname()
+            with open(whitelist_accession_f, 'wt') as f:
+                for accession in accessions:
+                    print(accession, file=f)
+
+            # Context-managerize eventually
+            taxonomy_tmp = tempfile.mkdtemp()
+            ncbitax.subset.subset_taxonomy(taxonomy, taxonomy_tmp, whitelist_accession_file=whitelist_accession_f)
+            shutil.move(taxonomy_tmp, taxonomy_dir)
+        else:
+            os.symlink(os.path.abspath(taxonomy), taxonomy_dir)
+    else:
+        if not taxonomy_exists:
+            raise FileNotFoundError('Taxonomy directory {} not found'.format(taxonomy_dir))
+        if args.subset_taxonomy:
+            raise KrakenUniqBuildError('Cannot subset taxonomy if already in db folder')
+
+    krakenuniq_tool = KrakenUniq()
+    options = {'--build': None}
+    if threads:
+        options['--threads'] = threads
+    if minimizer_len:
+        options['--minimizer-len'] = minimizer_len
+    if kmer_len:
+        options['--kmer-len'] = kmer_len
+    if max_db_size:
+        options['--max-db-size'] = max_db_size
+    if work_on_disk:
+        options['--work-on-disk'] = None
+    krakenuniq_tool.build(db, options=options)
+
+    if clean:
+        krakenuniq_tool.execute('krakenuniq-build', db, '', options={'--clean': None})
+
+
+
+
+def parser_krakenuniq(parser=argparse.ArgumentParser()):
+    parser.add_argument('db', help='Kraken database directory.')
+    parser.add_argument('in_bams', nargs='+', help='Input unaligned reads, BAM format.')
+    parser.add_argument('--outReports', dest='out_reports', nargs='+', help='Kraken summary report output file. Multiple filenames space separated.')
+    parser.add_argument('--outReads', dest='out_reads', nargs='+', help='Kraken per read classification output file. Multiple filenames space separated.')
+    parser.add_argument(
+        '--filterThreshold', dest='filter_threshold', default=0.05, type=float, help='Kraken filter threshold (default %(default)s)'
+    )
+    util.cmd.common_args(parser, (('threads', None), ('loglevel', None), ('version', None), ('tmp_dir', None)))
+    util.cmd.attach_main(parser, krakenuniq, split_args=True)
+    return parser
+def krakenuniq(db, in_bams, out_reports=None, out_reads=None, lock_memory=False, filter_threshold=None, threads=None):
+    '''
+        Classify reads by taxon using KrakenUniq
+    '''
+
+    assert out_reads or out_reports, ('Either --outReads or --outReport must be specified.')
+    kuniq_tool = KrakenUniq()
+    kuniq_tool.pipeline(db, in_bams, out_reports=out_reports, out_reads=out_reads,
+                        filter_threshold=filter_threshold, num_threads=threads)
+
+
+def fasta_library_accessions(library):
+    '''Parse accession from ids of fasta files in library directory. '''
+    library_accessions = set()
+    for dirpath, dirnames, filenames in os.walk(library, followlinks=True):
+        for filename in filenames:
+            if not filename.endswith('.fna') and not filename.endswith('.fa') and not filename.endswith('.ffn'):
+                continue
+            filepath = os.path.join(dirpath, filename)
+            for seqr in SeqIO.parse(filepath, 'fasta'):
+                name = seqr.name
+                # Search for accession
+                mo = re.search(r'([A-Z]+_?\d+\.\d+)', name)
+                if mo:
+                    accession = mo.group(1)
+                    library_accessions.add(accession)
+    return library_accessions
+
+
+
+def parser_kraken_taxlevel_summary(parser=argparse.ArgumentParser()):
+    parser.add_argument('summary_files_in', nargs="+", help='Kraken-format summary text file with tab-delimited taxonomic levels.')
+    parser.add_argument('--jsonOut', dest="json_out", type=argparse.FileType('w'), help='The path to a json file containing the relevant parsed summary data in json format.')
+    parser.add_argument('--csvOut', dest="csv_out", type=argparse.FileType('w'), help='The path to a csv file containing sample-specific counts.')
+    parser.add_argument('--taxHeading', nargs="+", dest="tax_headings", help='The taxonomic heading to analyze (default: %(default)s). More than one can be specified.', default="Viruses")
+    parser.add_argument('--taxlevelFocus', dest="taxlevel_focus", help='The taxonomic heading to summarize (totals by Genus, etc.) (default: %(default)s).', default="species")#,
+                        #choices=["species", "genus", "family", "order", "class", "phylum", "kingdom", "superkingdom"])
+    parser.add_argument('--topN', type=int, dest="top_n_entries", help='Only include the top N most abundant taxa by read count (default: %(default)s)', default=100)
+    parser.add_argument('--countThreshold', type=int, dest="count_threshold", help='Minimum number of reads to be included (default: %(default)s)', default=1)
+    parser.add_argument('--zeroFill', action='store_true', dest="zero_fill", help='When absent from a sample, write zeroes (rather than leaving blank).')
+    parser.add_argument('--noHist', action='store_true', dest="no_hist", help='Write out a report by-sample rather than a histogram.')
+    parser.add_argument('--includeRoot', action='store_true', dest="include_root", help='Include the count of reads at the root level and the unclassified bin.')
+    util.cmd.common_args(parser, (('loglevel', None), ('version', None), ('tmp_dir', None)))
+    util.cmd.attach_main(parser, taxlevel_summary, split_args=True)
+    return parser
+
+def taxlevel_summary(summary_files_in, json_out, csv_out, tax_headings, taxlevel_focus, top_n_entries, count_threshold, no_hist, zero_fill, include_root):
+    """
+        Aggregates taxonomic abundance data from multiple Kraken-format summary files.
+        It is intended to report information on a particular taxonomic level (--taxlevelFocus; ex. 'species'),
+        within a higher-level grouping (--taxHeading; ex. 'Viruses'). By default, when --taxHeading
+        is at the same level as --taxlevelFocus a summary with lines for each sample is emitted.
+        Otherwise, a histogram is returned. If per-sample information is desired, --noHist can be specified.
+        In per-sample data, the suffix "-pt" indicates percentage, so a value of 0.02 is 0.0002 of the total number of reads for the sample.
+        If --topN is specified, only the top N most abundant taxa are included in the histogram count or per-sample output.
+        If a number is specified for --countThreshold, only taxa with that number of reads (or greater) are included.
+        Full data returned via --jsonOut (filtered by --topN and --countThreshold), whereas -csvOut returns a summary.
+    """
+
+    samples = {}
+    same_level = False
+
+    Abundance = collections.namedtuple("Abundance", "percent,count,kmers,dup,cov")
+
+    def indent_len(in_string):
+        return len(in_string)-len(in_string.lstrip())
+
+    for f in list(summary_files_in):
+        sample_name, extension = os.path.splitext(f)
+        sample_summary = {}
+        sample_root_summary = {}
+        tax_headings_copy = [s.lower() for s in tax_headings]
+
+        # -----------------------------------------------------------------
+        # KrakenUniq has two lines prefixed by '#', a blank line,
+        # and then a TSV header beginning with "%". The column fields are:
+        # (NB:field names accurate, but space-separated in this comment
+        # for readability here)
+        #   %        reads  taxReads  kmers  dup   cov  taxID  rank          taxName
+        #   0.05591  2      0         13     1.85  NA   10239  superkingdom  Viruses
+        #
+        # Where the fields are:
+        #   %:
+        #   reads:
+        #   taxReads:
+        #   kmers: number of unique k-mers
+        #   dup: average number of times each unique k-mer has been seen
+        #   cov: coverage of the k-mers of the clade in the database
+        #   taxID:
+        #   rank: row["rank"]; A rank code (see list below)
+        #   taxName: row["sci_name"]; indented scientific name
+        #
+        # Taxonomic ranks used by KrakenUniq include:
+        #   unknown, no rank, sequence, assembly, subspecies,
+        #   species, species subgroup, species group, subgenus,
+        #   genus, tribe, subfamily, family, superfamily, parvorder,
+        #   infraorder, suborder, order, superorder, parvclass,
+        #   infraclass, subclass, class, superclass, subphylum,
+        #   phylum, kingdom, superkingdom, root
+        #
+        #   via: https://github.com/fbreitwieser/krakenuniq/blob/a8b4a2dbf50553e02d3cab3c32f93f91958aa575/src/taxdb.hpp#L96-L131
+        # -----------------------------------------------------------------
+        # Kraken (standard) reports lack header lines.
+        # (NB:field names below are only for reference. Space-separated for
+        # readability here)
+        #   %     reads  taxReads  rank  taxID      taxName
+        #   0.00  16     0         D     10239      Viruses
+        #
+        # Where the fields are:
+        #   %:        row["pct_of_reads"]; Percentage of reads covered by the clade rooted at this taxon
+        #   reads:    row["num_reads"]; Number of reads covered by the clade rooted at this taxon
+        #   taxReads: row["reads_exc_children"]; Number of reads assigned directly to this taxon
+        #   rank:     row["rank"]; A rank code, indicating (U)nclassified, (D)omain, (K)ingdom, (P)hylum, (C)lass, (O)rder, (F)amily, (G)enus, or (S)pecies. All other ranks are simply '-'.
+        #   taxID:    row["NCBI_tax_ID"]; NCBI taxonomy ID
+        #   taxName:  row["sci_name"]; indented scientific name
+        # -----------------------------------------------------------------
+
+
+        with util.file.open_or_gzopen(f, 'rU') as inf:
+            report_type=None
+            should_process = False
+            indent_of_selection = -1
+            currently_being_processed = ""
+            for lineno, line in enumerate(inf):
+                if len(line.rstrip('\r\n').strip()) == 0 or ( report_type != None and line.startswith("#") or line.startswith("%")):
+                    continue
+
+                # KrakenUniq is mentioned on the first line of
+                # summary reports created by KrakenUniq
+                if not report_type and "KrakenUniq" in line:
+                    report_type="krakenuniq"
+                    continue
+                elif not report_type:
+                    report_type="kraken"
+
+                csv.register_dialect('kraken_report', quoting=csv.QUOTE_MINIMAL, delimiter="\t")
+                if report_type == "kraken":
+                    fieldnames = [ "pct_of_reads",
+                                    "num_reads",
+                                    "reads_exc_children",
+                                    "rank",
+                                    "NCBI_tax_ID",
+                                    "sci_name"
+                                ]
+                elif report_type == "krakenuniq":
+                    fieldnames = [
+                                    "pct_of_reads",
+                                    "num_reads",
+                                    "reads_exc_children",
+                                    "uniq_kmers",
+                                    "kmer_dups",
+                                    "cov_of_clade_kmers",
+                                    "NCBI_tax_ID",
+                                    "rank",
+                                    "sci_name"
+                                ]
+                else:
+                    continue #never reached since we fall back to kraken above
+
+                row = next(csv.DictReader([line.strip().rstrip('\n')], fieldnames=fieldnames, dialect="kraken_report"))
+
+                try:
+                    indent_of_line = indent_len(row["sci_name"])
+                except AttributeError as e:
+                    log.warning("Report type: '{}'".format(report_type))
+                    log.warning("Issue with line {}: '{}'".format(lineno,line.strip().rstrip('\n')))
+                    log.warning("From file: {}".format(f))
+                # remove leading/trailing whitespace from each item
+                row = { k:v.strip() for k, v in row.items()}
+
+                # rows are formatted as described above.
+                # Kraken:
+                #   0.00  16  0   D   10239     Viruses
+                # KrakenUniq:
+                #   0.05591  2      0         13     1.85  NA   10239  superkingdom  Viruses
+
+                # if the root-level bins (root, unclassified) should be included, do so, but bypass normal
+                # stateful parsing logic since root does not have a distinct rank level
+                if row["sci_name"].lower() in ["root","unclassified"] and include_root:
+                    sample_root_summary[row["sci_name"]] = collections.OrderedDict()
+                    sample_root_summary[row["sci_name"]][row["sci_name"]] = Abundance(float(row["pct_of_reads"]), int(row["num_reads"]),row.get("kmers",None),row.get("dup",None),row.get("cov",None))
+                    continue
+
+                if indent_of_line <= indent_of_selection:
+                    should_process = False
+                    indent_of_selection=-1
+
+                if indent_of_selection == -1:
+                    if row["sci_name"].lower() in tax_headings_copy:
+                        tax_headings_copy.remove(row["sci_name"].lower())
+
+                        should_process = True
+                        indent_of_selection = indent_of_line
+                        currently_being_processed = row["sci_name"]
+                        sample_summary[currently_being_processed] = collections.OrderedDict()
+                        if row["rank"] == rank_code(taxlevel_focus) or row["rank"].lower().replace(" ","") == taxlevel_focus.lower().replace(" ",""):
+                            same_level = True
+                        if row["rank"] in ("-","no rank"):
+                            log.warning("Non-taxonomic parent level selected")
+
+                if should_process:
+                    # skip "-" rank levels since they do not occur at the sample level
+                    # otherwise include the taxon row if the rank matches the desired level of focus
+                    if (row["rank"] not in ("-","no rank") and (rank_code(taxlevel_focus) == row["rank"] or row["rank"].lower().replace(" ","") == taxlevel_focus.lower().replace(" ","")) ):
+                        if int(row["num_reads"])>=count_threshold:
+                            sample_summary[currently_being_processed][row["sci_name"]] = Abundance(float(row["pct_of_reads"]), int(row["num_reads"]),row.get("kmers",None),row.get("dup",None),row.get("cov",None))
+
+
+        for k,taxa in sample_summary.items():
+            sample_summary[k] = collections.OrderedDict(sorted(taxa.items(), key=lambda item: (item[1][1]) , reverse=True)[:top_n_entries])
+
+            if len(list(sample_summary[k].items()))>0:
+                log.info("{f}: most abundant among {heading} at the {level} level: "
+                            "\"{name}\" with {reads} reads ({percent:.2%} of total); "
+                            "included since >{threshold} read{plural}".format(
+                                                                          f=f,
+                                                                          heading=k,
+                                                                          level=taxlevel_focus,
+                                                                          name=list(sample_summary[k].items())[0][0],
+                                                                          reads=list(sample_summary[k].items())[0][1].count,
+                                                                          percent=list(sample_summary[k].items())[0][1].percent/100.0,
+                                                                          threshold=count_threshold,
+                                                                          plural="s" if count_threshold>1 else "" )
+                )
+
+        if include_root:
+            # include root-level bins (root, unclassified) in the returned data
+            for k,taxa in sample_root_summary.items():
+                assert (k not in sample_summary), "{k} already in sample summary".format(k=k)
+                sample_summary[k] = taxa
+        samples[sample_name] = sample_summary
+
+    if json_out != None:
+        json_summary = json.dumps(samples, sort_keys=True, indent=4, separators=(',', ': '))
+        json_out.write(json_summary)
+        json_out.close()
+
+
+    if csv_out != None:
+
+        # if we're writing out at the same level as the query header
+        # write out the fractions and counts
+        if same_level or no_hist:
+
+            fieldnames = set()
+            for sample, taxa in samples.items():
+                for heading,taxon in taxa.items():
+                    if len(taxon):
+                        for k in taxon.keys():
+                            fieldnames |= set([k+"-pt",k+"-ct"])
+
+            heading_columns = ["sample"]
+            if include_root:
+                root_fields = ["root-pt","root-ct","unclassified-pt","unclassified-ct"]
+                fieldnames -= set(root_fields)
+                heading_columns += root_fields
+
+            writer = csv.DictWriter(csv_out, restval=0 if zero_fill else '', fieldnames=heading_columns+sorted(list(fieldnames)))
+            writer.writeheader()
+
+            for sample, taxa in samples.items():
+                sample_dict = {}
+                sample_dict["sample"] = sample
+                for heading,taxon in taxa.items():
+                    for entry in taxon.keys():
+                        sample_dict[entry+"-pt"] = taxon[entry].percent
+                        sample_dict[entry+"-ct"] = taxon[entry].count
+                writer.writerow(sample_dict)
+
+
+            csv_out.close()
+
+        # otherwise write out a histogram
+        else:
+            count = 0
+            summary_counts = collections.defaultdict(dict)
+            for sample, totals in samples.items():
+                for heading,taxa in totals.items():
+                    for taxon in taxa.keys():
+                        if taxon not in summary_counts[heading].keys():
+                            summary_counts[heading][taxon] = 1
+                        else:
+                            summary_counts[heading][taxon] += 1
+
+            for k,taxa in summary_counts.items():
+                summary_counts[k] = collections.OrderedDict(sorted(taxa.items(), key=lambda item: (item[1]) , reverse=True))
+
+
+            fieldnames = ["heading","taxon","num_samples"]
+            writer = csv.DictWriter(csv_out, restval=0 if zero_fill else '', fieldnames=fieldnames)
+            writer.writeheader()
+
+            for heading,taxa_counts in summary_counts.items():
+                writer.writerows([{"heading":heading,"taxon":taxon,"num_samples":count} for taxon,count in taxa_counts.items()])
+
+            csv_out.close()
