+++ requirements-conda.txt
@@ -6,3 +6,4 @@ kaiju=1.6.3_yesimon
 krakenuniq=0.5.7_yesimon
 krona=2.7.1
 last=876
+ncbitax=0.1.2
