+++ setup.cfg
@@ -1,0 +1,5 @@
+[tool:pytest]
+addopts = --strict-markers
+markers =
+    slow: marks tests as slow (deselect with '-m "not slow"')
+    integration: marks tests as integration tests
