+++ test/test_metagenomics.py
@@ -1,9 +1,8 @@
-# Unit tests for metagenomics.py
-from builtins import super
-import six
+# Tests for metagenomics.py
 import argparse
 from collections import Counter
 import copy
+from io import StringIO
 import os.path
 from os.path import join
 import tempfile
@@ -13,50 +12,50 @@ import pytest
 
 import mock
 from mock import patch
+import ncbitax
 
 import tools.picard
-import metagenomics
 import util.file
 import util.misc
+import classify.blast
+import classify.metagenomics
+import classify.krona
+import classify.filtering
 from test import TestCaseWithTmp, assert_equal_bam_reads, _CPUS
 
-if six.PY2:
-    from StringIO import StringIO
-else:
-    from io import StringIO
 
 
-class TestCommandHelp(unittest.TestCase):
+# class TestCommandHelp(unittest.TestCase):
 
-    def test_help_parser_for_each_command(self):
-        for cmd_name, parser_fun in metagenomics.__commands__:
-            parser = parser_fun(argparse.ArgumentParser())
-            helpstring = parser.format_help()
+#     def test_help_parser_for_each_command(self):
+#         for cmd_name, parser_fun in classify.metagenomics.__commands__:
+#             parser = parser_fun(argparse.ArgumentParser())
+#             helpstring = parser.format_help()
 
 
-class TestKronaCalls(TestCaseWithTmp):
+# class TestKronaCalls(TestCaseWithTmp):
 
-    def setUp(self):
-        super().setUp()
-        patcher = patch('classify.krona.Krona', autospec=True)
-        self.addCleanup(patcher.stop)
-        self.mock_krona = patcher.start()
+#     def setUp(self):
+#         super().setUp()
+#         patcher = patch('classify.krona.Krona', autospec=True)
+#         self.addCleanup(patcher.stop)
+#         self.mock_krona = patcher.start()
 
-        self.inTsv = util.file.mkstempfname('.tsv')
-        self.db = tempfile.mkdtemp('db')
+#         self.inTsv = util.file.mkstempfname('.tsv')
+#         self.db = tempfile.mkdtemp('db')
 
-    def test_krona_import_taxonomy(self):
-        out_html = util.file.mkstempfname('.html')
-        metagenomics.krona(self.inTsv, self.db, out_html, queryColumn=3, taxidColumn=5, scoreColumn=7,
-                           noHits=True, noRank=True, inputType='tsv')
-        self.mock_krona().import_taxonomy.assert_called_once_with(
-            self.db, [self.inTsv + ',' + os.path.basename(self.inTsv)], out_html, query_column=3, taxid_column=5, score_column=7,
-            no_hits=True, no_rank=True, magnitude_column=None)
+    # def test_krona_import_taxonomy(self):
+    #     out_html = util.file.mkstempfname('.html')
+    #     classify.krona.krona(self.inTsv, self.db, out_html, queryColumn=3, taxidColumn=5, scoreColumn=7,
+    #                        noHits=True, noRank=True, inputType='tsv')
+    #     self.mock_krona().import_taxonomy.assert_called_once_with(
+    #         self.db, [self.inTsv + ',' + os.path.basename(self.inTsv)], out_html, query_column=3, taxid_column=5, score_column=7,
+    #         no_hits=True, no_rank=True, magnitude_column=None)
 
 
 @pytest.fixture
 def taxa_db_simple():
-    db = metagenomics.TaxonomyDb()
+    db = ncbitax.TaxonomyDb()
     db.gis = {1:2, 2:3, 3:4, 4:5}
     db.parents = {1: 1, 2: 1, 3: 2, 4: 3, 5: 4}
     return db
@@ -64,7 +63,7 @@ def taxa_db_simple():
 
 @pytest.fixture
 def taxa_db(parents, names, ranks):
-    db = metagenomics.TaxonomyDb()
+    db = ncbitax.TaxonomyDb()
     db.parents = parents
     db.names = names
     db.ranks = ranks
@@ -130,50 +129,38 @@ def simple_m8():
 
 def test_tree_level_lookup(parents):
     level_cache = {1: 1}
-    assert metagenomics.tree_level_lookup(parents, 1, level_cache) == 1
-    assert metagenomics.tree_level_lookup(parents, 3, level_cache) == 2
-    assert metagenomics.tree_level_lookup(parents, 12, level_cache) == 6
+    assert classify.metagenomics.tree_level_lookup(parents, 1, level_cache) == 1
+    assert classify.metagenomics.tree_level_lookup(parents, 3, level_cache) == 2
+    assert classify.metagenomics.tree_level_lookup(parents, 12, level_cache) == 6
     level_cache = {1: 1}
-    assert metagenomics.tree_level_lookup(parents, 12, level_cache) == 6
-    assert metagenomics.tree_level_lookup(parents, 8, level_cache) == 5
+    assert classify.metagenomics.tree_level_lookup(parents, 12, level_cache) == 6
+    assert classify.metagenomics.tree_level_lookup(parents, 8, level_cache) == 5
 
 
 def test_push_up_tree_hits(parents):
     hits = Counter({1: 3, 3: 5, 6: 3, 7: 3, 13: 5})
     with pytest.raises(AssertionError):
-        metagenomics.push_up_tree_hits(parents, hits)
+        classify.metagenomics.push_up_tree_hits(parents, hits)
 
     expected = hits.copy()
-    assert metagenomics.push_up_tree_hits(parents, hits.copy(), min_support=1) == expected
+    assert classify.metagenomics.push_up_tree_hits(parents, hits.copy(), min_support=1) == expected
 
     expected = Counter({3: 5, 6: 6, 13: 5})
-    assert metagenomics.push_up_tree_hits(parents, hits.copy(), min_support=5) == expected
+    assert classify.metagenomics.push_up_tree_hits(parents, hits.copy(), min_support=5) == expected
 
-    assert (metagenomics.push_up_tree_hits(parents, hits.copy(), min_support=10) ==
+    assert (classify.metagenomics.push_up_tree_hits(parents, hits.copy(), min_support=10) ==
             Counter({6: 11}))
-    assert (metagenomics.push_up_tree_hits(parents, hits.copy(), min_support=18) ==
+    assert (classify.metagenomics.push_up_tree_hits(parents, hits.copy(), min_support=18) ==
             Counter({1: 19}))
-    assert (metagenomics.push_up_tree_hits(parents, hits.copy(), min_support_percent=100) ==
+    assert (classify.metagenomics.push_up_tree_hits(parents, hits.copy(), min_support_percent=100) ==
             Counter({1: 19}))
-
-
-def test_parents_to_children(parents):
-    children = metagenomics.parents_to_children(parents)
-    assert children[1] == [3]
-
-
-def test_rank_code():
-    assert metagenomics.rank_code('species') == 'S'
-    assert metagenomics.rank_code('genus') == 'G'
-    assert metagenomics.rank_code('superkingdom') == 'D'
-    assert metagenomics.rank_code('nonexist') == '-'
 
 
 def test_blast_records(simple_m8):
     test_path = join(util.file.get_test_input_path(),
                              'TestTaxonomy')
     with simple_m8 as f:
-        records = list(metagenomics.blast_records(f))
+        records = list(classify.blast.blast_records(f))
     assert len(records) == 110
     assert records[0].bit_score == 63.5
     assert records[-1].bit_score == 67.4
@@ -193,7 +180,7 @@ def test_blast_lca(taxa_db_simple, simple_m8):
     """)
     out = StringIO()
     with simple_m8 as f:
-        metagenomics.blast_lca(taxa_db_simple, f, out, paired=True)
+        classify.blast.blast_lca(taxa_db_simple, f, out, paired=True)
         out.seek(0)
         assert out.read() == expected
 
@@ -202,33 +189,33 @@ def test_paired_query_id():
     tup = ['query', 'gi|10|else', 90., 80, 60, 2, 30, 80,
            1100, 1150, 1e-7, 64.5, []]
 
-    blast1 = metagenomics.BlastRecord(*tup)
-    assert metagenomics.paired_query_id(blast1) == blast1
+    blast1 = classify.blast.BlastRecord(*tup)
+    assert classify.blast.paired_query_id(blast1) == blast1
 
     new_tup = copy.copy(tup)
     new_tup[0] = 'query/1'
-    new_blast1 = metagenomics.BlastRecord(*new_tup)
-    assert metagenomics.paired_query_id(new_blast1) == blast1
+    new_blast1 = classify.blast.BlastRecord(*new_tup)
+    assert classify.blast.paired_query_id(new_blast1) == blast1
 
     new_tup = copy.copy(tup)
     new_tup[0] = 'query/2'
-    new_blast1 = metagenomics.BlastRecord(*new_tup)
-    assert metagenomics.paired_query_id(new_blast1) == blast1
+    new_blast1 = classify.blast.BlastRecord(*new_tup)
+    assert classify.blast.paired_query_id(new_blast1) == blast1
 
     new_tup = copy.copy(tup)
     new_tup[0] = 'query/3'
-    new_blast1 = metagenomics.BlastRecord(*new_tup)
-    assert metagenomics.paired_query_id(new_blast1) == new_blast1
+    new_blast1 = classify.blast.BlastRecord(*new_tup)
+    assert classify.blast.paired_query_id(new_blast1) == new_blast1
 
 
 def test_translate_gi_to_tax_id(taxa_db_simple):
     tup = ['query', 'gi|4|else', 90., 80, 60, 2, 30, 80,
            1100, 1150, 1e-7, 64.5, []]
-    blast1 = metagenomics.BlastRecord(*tup)
+    blast1 = classify.blast.BlastRecord(*tup)
 
     tup[1] = 5
-    expected = metagenomics.BlastRecord(*tup)
-    assert metagenomics.translate_gi_to_tax_id(taxa_db_simple, blast1) == expected
+    expected = classify.blast.BlastRecord(*tup)
+    assert classify.blast.translate_gi_to_tax_id(taxa_db_simple, blast1) == expected
 
 
 def test_kraken_dfs_report(taxa_db):
@@ -243,17 +230,9 @@ def test_kraken_dfs_report(taxa_db):
     25.72\t107\t0\t-\t8\t        eight
     25.72\t107\t107\tS\t12\t          twelve
     ''')
-    report = metagenomics.kraken_dfs_report(taxa_db, hits)
+    report = classify.metagenomics.kraken_dfs_report(taxa_db, hits)
     text_report = '\n'.join(list(report)) + '\n'
     assert text_report == expected
-
-
-def test_coverage_lca(taxa_db):
-    assert metagenomics.coverage_lca([10, 11, 12], taxa_db.parents) == 6
-    assert metagenomics.coverage_lca([1, 3], taxa_db.parents) == 1
-    assert metagenomics.coverage_lca([6, 7, 8], taxa_db.parents) == 6
-    assert metagenomics.coverage_lca([10, 11, 12], taxa_db.parents, 50) == 7
-    assert metagenomics.coverage_lca([9], taxa_db.parents) is None
 
 
 def test_krakenuniq(mocker):
@@ -264,7 +243,7 @@ def test_krakenuniq(mocker):
         '--outReports', 'output.report',
         '--outReads', 'output.reads',
     ]
-    args = metagenomics.parser_krakenuniq(argparse.ArgumentParser()).parse_args(args)
+    args = classify.kraken.parser_krakenuniq(argparse.ArgumentParser()).parse_args(args)
     args.func_main(args)
     p.assert_called_with('db', ['input.bam'], num_threads=mock.ANY, filter_threshold=mock.ANY, out_reports=['output.report'], out_reads=['output.reads'])
 
@@ -278,7 +257,7 @@ def test_kaiju(mocker):
         'output.report',
         '--outReads', 'output.reads',
     ]
-    args = metagenomics.parser_kaiju(argparse.ArgumentParser()).parse_args(args)
+    args = classify.kaiju.parser_kaiju(argparse.ArgumentParser()).parse_args(args)
     args.func_main(args)
     p.assert_called_with('db.fmi', 'tax_db', 'input.bam', output_report='output.report', num_threads=mock.ANY, output_reads='output.reads')
 
@@ -299,7 +278,7 @@ class TestBamFilter(TestCaseWithTmp):
             "--taxNames",
             "Ebolavirus"
         ]
-        args = metagenomics.parser_filter_bam_to_taxa(argparse.ArgumentParser()).parse_args(args)
+        args = classify.filtering.parser_filter_bam_to_taxa(argparse.ArgumentParser()).parse_args(args)
         args.func_main(args)
 
         expected_bam = os.path.join(input_dir,"expected.bam")
@@ -320,7 +299,7 @@ class TestBamFilter(TestCaseWithTmp):
             "--taxIDs",
             "186538"
         ]
-        args = metagenomics.parser_filter_bam_to_taxa(argparse.ArgumentParser()).parse_args(args)
+        args = classify.filtering.parser_filter_bam_to_taxa(argparse.ArgumentParser()).parse_args(args)
         args.func_main(args)
 
         expected_bam = os.path.join(input_dir,"expected.bam")
